daos: abstract to mfu_daos_util
Abstracted daos-specifc functions to new files:
- mfu_daos.h
- mfu_daos.c

cmake will now conditionally include these files.
