+++ src/common/mfu_daos.h
@@ -1,0 +1,52 @@
+#include "mfu.h"
+
+#include <daos.h>
+
+enum handleType {
+    POOL_HANDLE,
+    CONT_HANDLE,
+    ARRAY_HANDLE
+};
+
+/* struct for holding DAOS arguments */
+typedef struct {
+    daos_handle_t src_poh; /* source pool handle */
+    daos_handle_t dst_poh; /* destination pool handle */
+    daos_handle_t src_coh; /* source container handle */
+    daos_handle_t dst_coh; /* destination container handle */
+    uuid_t src_pool_uuid;  /* source pool UUID */
+    uuid_t dst_pool_uuid;  /* destination pool UUID */
+    uuid_t src_cont_uuid;  /* source container UUID */
+    uuid_t dst_cont_uuid;  /* destination container UUID */
+    char* dfs_prefix;      /* prefix for UNS */
+} daos_args_t;
+
+/* Return a newly allocated daos_args_t structure.
+ * Set default values on its fields. */
+daos_args_t* daos_args_new(void);
+
+/* free a daos_args_t structure */
+void daos_args_delete(daos_args_t** pda);
+
+/* Setup DAOS arguments.
+ * Connect to pools.
+ * Open containers.
+ * Mount DFS. 
+ * Returns 1 on error, 0 on success */
+int daos_setup(
+  int rank,
+  char** argpaths,
+  daos_args_t* da,
+  mfu_file_t* mfu_src_file,
+  mfu_file_t* mfu_dst_file
+);
+
+/* Unmount DFS.
+ * Disconnect from pool/cont.
+ * Cleanup DAOS-related vars, handles. 
+ * Finalize DAOS. */
+int daos_cleanup(
+  daos_args_t* da,
+  mfu_file_t* mfu_src_file,
+  mfu_file_t* mfu_dst_file
+);
