+++ src/common/mfu_util.c
@@ -25,11 +25,6 @@
 #include <lustre/lustreapi.h>
 #endif
 
-#ifdef DAOS_SUPPORT
-#include <gurt/common.h>
-#include <gurt/hash.h>
-#endif
-
 int mfu_initialized = 0;
 
 /* set globals */
@@ -41,193 +36,6 @@ mfu_loglevel mfu_debug_level = MFU_LOG_ERR;
 
 /* default progress message timeout in seconds */
 int mfu_progress_timeout = 10;
-
-/***** DAOS utility functions ******/
-#ifdef DAOS_SUPPORT
-bool daos_uuid_valid(const uuid_t uuid)
-{
-    return uuid && !uuid_is_null(uuid);
-}
-
-/* Distribute process 0's pool or container handle to others. */
-void daos_bcast_handle(
-  int rank,              /* root rank for broadcast */
-  daos_handle_t* handle, /* handle value to be broadcasted */
-  daos_handle_t* poh,    /* daos pool for global2local conversion of container handle */
-  enum handleType type)  /* handle type: POOL_HANDLE or CONT_HANDLE */
-{
-    int rc;
-
-    d_iov_t global;
-    global.iov_buf     = NULL;
-    global.iov_buf_len = 0;
-    global.iov_len     = 0;
-
-    /* Get the global handle size. */
-    if (rank == 0) {
-        if (type == POOL_HANDLE) {
-            rc = daos_pool_local2global(*handle, &global);
-        } else {
-            rc = daos_cont_local2global(*handle, &global);
-        }
-        if (rc != 0) {
-            MFU_ABORT(-1, "Failed to get global handle size");
-        }
-    }
-
-    /* broadcast size of global handle */
-    MPI_Bcast(&global.iov_buf_len, 1, MPI_UINT64_T, 0, MPI_COMM_WORLD);
-
-    /* allocate memory to hold global handle value */
-    global.iov_len = global.iov_buf_len;
-    global.iov_buf = MFU_MALLOC(global.iov_buf_len);
-
-    /* convert from local handle to global handle */
-    if (rank == 0) {
-       if (type == POOL_HANDLE) {
-           rc = daos_pool_local2global(*handle, &global);
-       } else {
-           rc = daos_cont_local2global(*handle, &global);
-       }
-       if (rc != 0) {
-           MFU_ABORT(-1, "Failed to create global handle");
-       }
-    }
-
-    /* broadcast global handle value */
-    MPI_Bcast(global.iov_buf, global.iov_buf_len, MPI_BYTE, 0, MPI_COMM_WORLD);
-
-    /* convert global handle to local value */
-    if (rank != 0) {
-        if (type == POOL_HANDLE) {
-            rc = daos_pool_global2local(global, handle);
-        } else {
-            rc = daos_cont_global2local(*poh, global, handle);
-        }
-        if (rc != 0) {
-            MFU_ABORT(-1, "Failed to get local handle");
-        }
-    }
-
-    /* free temporary buffer used to hold global handle value */
-    mfu_free(&global.iov_buf);
-}
-
-int daos_connect(
-  int rank,
-  uuid_t pool_uuid,
-  uuid_t cont_uuid,
-  daos_handle_t* poh,
-  daos_handle_t* coh,
-  bool connect_pool,
-  bool create_cont)
-{
-    /* assume failure until otherwise */
-    int valid = 0;
-    int rc;
-
-    /* have rank 0 connect to the pool and container,
-     * we'll then broadcast the handle ids from rank 0 to everyone else */
-    if (rank == 0) {
-        /* Connect to DAOS pool */
-        if (connect_pool) {
-            daos_pool_info_t pool_info = {0};
-#if DAOS_API_VERSION_MAJOR < 1
-            rc = daos_pool_connect(pool_uuid, NULL, NULL, DAOS_PC_RW,
-                    poh, &pool_info, NULL);
-#else
-            rc = daos_pool_connect(pool_uuid, NULL, DAOS_PC_RW,
-                    poh, &pool_info, NULL);
-#endif
-            if (rc != 0) {
-                MFU_LOG(MFU_LOG_ERR, "Failed to connect to pool");
-                goto bcast;
-            }
-        }
-
-        /* Try to open the container 
-         * If NOEXIST we create it */
-        daos_cont_info_t co_info = {0};
-        rc = daos_cont_open(*poh, cont_uuid, DAOS_COO_RW, coh, &co_info, NULL);
-        if (rc != 0) {
-            if (!create_cont) {
-                MFU_LOG(MFU_LOG_ERR, "Failed to open DFS container");
-                goto bcast;
-            }
-            
-            rc = dfs_cont_create(*poh, cont_uuid, NULL, NULL, NULL);
-            if (rc != 0) {
-                MFU_LOG(MFU_LOG_ERR, "Failed to create DFS container");
-                goto bcast;
-            }
-
-            /* try to open it again */
-            rc = daos_cont_open(*poh, cont_uuid, DAOS_COO_RW, coh, &co_info, NULL);
-            if (rc != 0) {
-                MFU_LOG(MFU_LOG_ERR, "Failed to open DFS container");
-                goto bcast;
-            }
-        }
-
-        /* everything looks good so far */
-        valid = 1;
-    }
-
-bcast:
-    /* broadcast valid from rank 0 */
-    MPI_Bcast(&valid, 1, MPI_INT, 0, MPI_COMM_WORLD);
-
-    /* return if invalid */
-    if (valid == 0) {
-        return -1;
-    }
-
-    /* broadcast pool handle from rank 0 
-     * If connect_pool is false, then the handle was unchanged */
-    if (connect_pool) {
-        daos_bcast_handle(rank, poh, poh, POOL_HANDLE);
-    }
-
-    /* broadcast container handle from rank 0 */
-    daos_bcast_handle(rank, coh, poh, CONT_HANDLE);
-
-    return 0;
-}
-
-int daos_mount(
-  mfu_file_t* mfu_file,
-  daos_handle_t* poh,
-  daos_handle_t* coh)
-{
-    /* Mount dfs */
-    int rc = dfs_mount(*poh, *coh, O_RDWR, &mfu_file->dfs);
-    if (rc != 0) {
-        MFU_LOG(MFU_LOG_ERR, "Failed to mount DAOS filesystem (DFS): "
-                MFU_ERRF, MFU_ERRP(-MFU_ERR_DAOS));
-        rc = -1;
-    }
-
-    return rc;
-}
-
-int daos_umount(
-  mfu_file_t* mfu_file)
-{
-    /* Unmount dfs */
-    int rc = dfs_umount(mfu_file->dfs);
-    if (rc != 0) {
-        MFU_LOG(MFU_LOG_ERR, "Failed to unmount DFS namespace");
-        rc = -1;
-    }
-
-    /* Clean up the hash */
-    if (mfu_file->dfs_hash != NULL) {
-        d_hash_table_destroy(mfu_file->dfs_hash, true);
-    }
-
-    return rc;
-}
-#endif
 
 /* initialize mfu library,
  * reference counting allows for multiple init/finalize pairs */
